+++ zaifbot/bot_common/__init__.py
@@ -1,10 +1,1 @@
-from os import environ as env
 
-
-def preset_keys(key, secret):
-    env['ZAIFBOT_KEY'] = key
-    env['ZAIFBOT_SECRET'] = secret
-
-
-def get_keys():
-    return env['ZAIFBOT_KEY'], env['ZAIFBOT_SECRET']
